moved the analysis to the pass
