+++ vyper/venom/analysis/__init__.py
@@ -6,6 +6,5 @@ from .fcg import FCGAnalysis
 from .liveness import LivenessAnalysis
 from .mem_alias import MemoryAliasAnalysis
 from .mem_ssa import MemSSA
-from .phi_reach import PhiReachingAnalysis
 from .reachable import ReachableAnalysis
 from .var_definition import VarDefinition
